requirements.txt
crimson-intelli-type==0.4.0
pyyaml
